Fix bug - list deletion without beeing the author

+++ src/server/routes/List.ts
@@ -303,6 +303,10 @@ List.delete("/",
 		const list = await prisma.list.findUnique({ where: { id } });
 		if (!list)
 			return res.status(404).json({ success: false, message: "List not found" });
+		// Check if user is member of list
+		const author =  await prisma.list.findFirst({ where: { id, subscribers: { some: { id: getTokenId(token) } } } });
+		if (!author)
+			return res.status(406).json({ success: false, message: `You are not member of ${list.title}(${list.id}). \nPlease ask author of this Todo-list to add you` });
 		// Delete list
 		await prisma.list.delete({ where: { id } });
 		res.status(200).json({ success: true, message: list });
